+++ packages/guide/src/app/locations/components/shops/generic-shop/generic-shop.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'app-generic-shop',
+    templateUrl: './generic-shop.component.html',
+})
+export class GenericShopComponent {
+}