+++ packages/guide/src/app/locations/types/festival-route-data.type.ts
@@ -1,0 +1,5 @@
+import { FestivalName } from "@ci/data-types";
+
+export type FestivalRouteData = {
+    name: FestivalName,
+}