+++ packages/pak-files-parser/src/types/object-path.type.ts
@@ -1,0 +1,4 @@
+export type ObjectPath = {
+    "ObjectName": string
+    "ObjectPath": string
+}